refactor: remove vitest.config.ts and migrate to vitest.config.mts with updated configuration
