+++ packages/@textlint/markdown-to-ast/test/parsing-test.ts
@@ -4,7 +4,7 @@ import { describe, it } from "vitest";
 import fs from "node:fs";
 import path from "node:path";
 import { test as astTest } from "@textlint/ast-tester";
-import { parse } from "../lib/src/index.js";
+import { parse } from "../src/index.js";
 
 describe("parsing", function () {
     const fixtureDir = path.join(__dirname, "fixtures");
