+++ packages/textlint-tester/tsconfig.json
@@ -1,8 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "lib",
-    "types": ["node", "mocha"]
+    "outDir": "lib"
   },
   "include": [
     "src/**/*"
