+++ packages/textlint/test/createLinter/fixtures/modules/textlint-rule-cjs/index.cjs
@@ -1,11 +1,11 @@
 /**
  * @type {import("@textlint/types").TextlintRuleModule}}}
  */
-module.exports = (context) => {
+export default (context) => {
     const { Syntax, RuleError } = context;
     return {
         [Syntax.Str](node) {
             context.report(node, new RuleError("CJS"));
         }
     };
-};
+}
